IActionFilter HomeWork

+++ Booking_Ex/Controllers/CategoryController.cs
@@ -2,6 +2,7 @@ using System.Security.Claims;
 using ASP_.Net_Core_Class_Home_Work.Data.DAL;
 using ASP_.Net_Core_Class_Home_Work.Data.Entities;
 using ASP_.Net_Core_Class_Home_Work.Middleware;
+using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.Extensions.Primitives;
 
 namespace ASP_.Net_Core_Class_Home_Work.Controllers;
@@ -9,40 +10,49 @@ using Microsoft.AspNetCore.Mvc;
 
 [Route("api/category")]
 [ApiController]
-public class CategoryController : ControllerBase
+// IActionFilter - засоби для дій що виконуються до вільного Action-iв
+public class CategoryController : BackendController
 {
     private readonly DataAccessor _DataAccessor;
-
-    public CategoryController(DataAccessor dataAccessor)
+    private readonly ILogger<CategoryController> _logger;
+    // private bool isAuthentication;
+    // private bool iaAdmin;
+    public CategoryController(DataAccessor dataAccessor, ILogger<CategoryController> _logger)
     {
         _DataAccessor = dataAccessor;
+        this._logger = _logger;
     }
-
+    // метод IActionFilter, що виконується ДО дій контролера (DoGet, DoPost,...)
+    // [NonAction]
+    // public void OnActionExecuting(ActionExecutingContext context)
+    // {
+    //    
+    //     // У проекті є жві авторизаціі - через сесіі та через токени
+    //     // Первинна авторизація за сесією 
+    //     // Дані авторизаціі за токеном шукаємо за типом авторизації яку ми встановили як назва класу AuthTokenMiddleware
+    //     var identity = User.Identities.FirstOrDefault(i => i.AuthenticationType == nameof(AuthSessionMiddleware));
+    //     identity ??=  User.Identities.FirstOrDefault(i => i.AuthenticationType == nameof(AuthTokenMiddleware));
+    //     this.isAuthentication = identity != null;
+    //     String? useRole = identity?.Claims.FirstOrDefault(c=>c.Type==ClaimTypes.Role)?.Value; 
+    //     this.iaAdmin = "Admin".Equals(useRole);
+    //    
+    // }
+    
     [HttpGet]
     public List<Category> DoGet()
     {
-        String? useRole = HttpContext.User.Claims.FirstOrDefault(c=>c.Type==ClaimTypes.Role)?.Value;
-        bool iaAdmin = "Admin".Equals(useRole);
        return _DataAccessor._ContentDao.GetCategories(includeDelete: iaAdmin);
     }
 
     [HttpPost]
     public string DoPost([FromForm]CategoryPostModel model)
     {
-        // У проекті є жві авторизаціі - через сесіі та через токени
-        // Первинна авторизація за сесією 
-        // Дані авторизаціі за токеном шукаємо за типом авторизації яку ми встановили як назва класу AuthTokenMiddleware
-        var identity= User.Identities.FirstOrDefault(i => i.AuthenticationType == nameof(AuthTokenMiddleware));
-        if (identity == null)
+       
+        // var identity = User.Identities.FirstOrDefault(i => i.AuthenticationType == nameof(AuthSessionMiddleware));
+        // identity ??=  User.Identities.FirstOrDefault(i => i.AuthenticationType == nameof(AuthTokenMiddleware));
+        if (GetAdminAuthMessage() is String msg)
         {
-            // якщо авторизація не пройдена то повідомлення а Items
-            Response.StatusCode = StatusCodes.Status401Unauthorized;
-            return HttpContext.Items[nameof(AuthTokenMiddleware)]?.ToString() ?? "";
-        }
-        if ( identity.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role).Value != "Admin")
-        {
-            Response.StatusCode = StatusCodes.Status403Forbidden;
-            return "Access to API forbidden!";
+           return msg;
         }
         try
         {
@@ -76,9 +86,111 @@ public class CategoryController : ControllerBase
         }
     }
 
+    [HttpPut] //Update
+    public string DoPut([FromForm] CategoryPostModel model)
+    {
+        if (GetAdminAuthMessage() is String msg)
+        {
+            return msg;
+        }
+        // перевіряємо CategoryId на наявність
+        if (model.CategoryId==null || model.CategoryId==default(Guid))
+        {
+            Response.StatusCode = StatusCodes.Status422UnprocessableEntity;
+            return "Missing required parameter: 'category-id' ";
+        }
+        // перевіряємо чи є така ктегорія
+        Category? category = _DataAccessor._ContentDao.GetCategoryById(model.CategoryId.Value);
+        if (category == null)
+        {
+            Response.StatusCode = StatusCodes.Status422UnprocessableEntity;
+            return $"Parameter 'category-id' ({model.CategoryId.Value}) belongs to no entity. ";
+        }
+        // оновлення даних - якщо немає данних то залішається попереднє значення
+        if (!string.IsNullOrEmpty(model.Name))
+        {
+            category.Name = model.Name;
+        }
+        if (!string.IsNullOrEmpty(model.Description))
+        {
+            category.Description = model.Description;
+        }
+        if (!string.IsNullOrEmpty(model.Slug))
+        {
+            category.Slug = model.Slug;
+        }
+        if (model.Photo != null)   // передається новий файл - зберігаємо новий, видаляємо старий
+        {
+
+            try
+
+            {
+
+                String? fileName = null;
+
+                String ext = Path.GetExtension(model.Photo.FileName);
+
+                String path = Directory.GetCurrentDirectory() + "/wwwroot/img/content/";
+
+                String pathName;
+
+                do
+
+                {
+
+                    fileName = Guid.NewGuid() + ext;
+
+                    pathName = path + fileName;
+
+                }
+
+                while (System.IO.File.Exists(pathName));
+
+                using var stream = System.IO.File.OpenWrite(pathName);
+
+                model.Photo.CopyTo(stream);
+                // новий файл успішно завантажений - видаляємо старий
+
+                if (!String.IsNullOrEmpty((category.PhotoUrl)))
+                {
+                    try
+                    {
+                        System.IO.File.Delete(path + category.PhotoUrl);
+                    }
+                    catch
+                    {
+                        _logger.LogWarning(category.PhotoUrl+"not deleted");
+                    }
+                     
+                } 
+                // зберігаємо нове ім'я 
+
+                category.PhotoUrl = fileName;
+
+            }
+            catch (Exception ex)
+
+            {
+                _logger.LogWarning(ex.Message);
+                Response.StatusCode = StatusCodes.Status400BadRequest;
+
+                return "Error uploading file";
+
+            }
+
+        }
+        _DataAccessor._ContentDao.UpdateCategory(category);
+        Response.StatusCode = StatusCodes.Status200OK;
+        return "Update";
+    }
+
     [HttpDelete("{id}")]
     public string DoDelete(Guid id)
     {
+        if (GetAdminAuthMessage() is String msg)
+        {
+            return msg;
+        }
         _DataAccessor._ContentDao.DeleteCategory(id);
         Response.StatusCode = StatusCodes.Status202Accepted;
         return "Ok";
@@ -100,6 +212,10 @@ public class CategoryController : ControllerBase
    // Другій НЕ позначений метод private щоб не було конфлікту
     private string DoRestore()
     {
+        if (GetAdminAuthMessage() is String msg)
+        {
+            return msg;
+        }
         string? id = Request.Query["id"].FirstOrDefault();
         try
         {
@@ -114,7 +230,13 @@ public class CategoryController : ControllerBase
         Response.StatusCode = StatusCodes.Status202Accepted;
         return "RESTORE works with id = " + id;
     }
-
+ 
+    // метод IActionFilter, що виконується ПІСЛЯ дій контролера (DoGet, DoPost,...)
+    // [NonAction] // Якщо неможна зробити метод private то позначаємо атрибутом [NonAction]
+    // public void OnActionExecuted(ActionExecutedContext context)
+    // {
+    //    
+    // }
     public class CategoryPostModel
     {
         [FromForm(Name="category-name")]
@@ -128,5 +250,9 @@ public class CategoryController : ControllerBase
         
         [FromForm(Name="category-photo")]
         public IFormFile? Photo { set; get; }
+        
+        [FromForm(Name="category-id")]
+        public Guid? CategoryId { set; get; }
+
     }
 }
